requirements.txt
anthropic>=0.42.0
httpx>=0.27.0
python-dotenv>=1.0.0
fpdf2>=2.8.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
